Correctly parse row changesets
Although I didn't find any documentation for this,
mork files seem to be containing groups (aka changesets)
with rows that are not enclosed by a table. These rows
seem to represent updates to existing rows. the row to
apply the update to is identified by the row id and scope.
According to
https://developer.mozilla.org/en-US/docs/Mozilla/Tech/Mork#Rows,
every row id is unique per scope, even across tables.