Move the forceUpdateTimer to the UnreadMonitor thread
Previously, the forceUpdateTimer was executed on the main thread.
Because QTimers can only be started and stopped on the thread they
belong to, this meant that changing the settings about the force unread
timer did not apply, because the `slotSettingsChanged` function is
called on the UnreadMonitor tread. So editing the timer did just print
the following error to the log:
QObject::startTimer: Timers cannot be started from another thread

The setting would only really apply when Birdtray was restarted.

The fix is to let the mForceUpdateTimer be a child of UnreadMonitor.
This causes the timer to be moved to the UnreadMonitor thread by the
`moveToThread(this)` call in the constructor. The initial start of the
timer is then moved to the run method, so it happens on the thread.