Add a setting to change the searched Thunderbird process name on Windows
This setting is only useful on Windows and allows to change the name of
the process that is searched to detect if Thunderbird is running.
Changing the setting will take effect when saving, except for if
Birdtray already found a process with the old settings.
In that case, Birdtray has to be restarted or the currently tracked
process has to exit.